+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.6] 2022-05-28
+## [0.6.6] 2022-05-30
 
 ### Added
+
+- Drag & drop functionality was added for scene hierarchy's window.
 
 ### Changed